* Add missing qspi to the Pi to be able to program bitstream faster. Done
* used IS25LP016D-JBLE as flash, but
 it has problems with enabling QE for QSPI.
Relace with the model used on icebreaker
This is already done, comment should be removed

* ecp5 SPI clock pin can't be used as an input
 so have to [add a bodge wire](https://github.com/mattvenn/basic-ecp5-pcb/issues/3)
for SPI peripheral comms with Pi.
See the test-spi-clk-in branch for a workaround.
Worked around by connecting this pin to a second FPGA pin

* connected I2C pins are just GPIOs. I have found the Pi's I2C hardware unusable, so always use the slower bitbanged version.
Connected I2C pins to official pins.
Removed pull resistors as this is not needed